requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
google-cloud-firestore